feat: Install cobra-cli
